DataBaseException adicionado

+++ src/main/java/com/beawata/petcare/services/EspecieService.java
@@ -3,11 +3,14 @@ package com.beawata.petcare.services;
 import com.beawata.petcare.dto.EspecieDTO;
 import com.beawata.petcare.entities.Especie;
 import com.beawata.petcare.repositories.EspecieRepository;
+import com.beawata.petcare.services.exceptions.DataBaseException;
 import com.beawata.petcare.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -43,17 +46,30 @@ public class EspecieService {
     //Método PUT para atualizar uma especie
     @Transactional
     public EspecieDTO update(EspecieDTO dto){
-        Especie entity = especieRepository.findById(dto.getId())
-                .orElseThrow(() -> new ResourceNotFoundException("Espécie não encontrada"));
-        copyDtoToEntity(dto, entity);
-        entity = especieRepository.save(entity);
-        return new EspecieDTO(entity);
+        try {
+            Especie entity = especieRepository.findById(dto.getId())
+                    .orElseThrow(() -> new ResourceNotFoundException("Espécie não encontrada"));
+            copyDtoToEntity(dto, entity);
+            entity = especieRepository.save(entity);
+            return new EspecieDTO(entity);
+        }
+        catch (ResourceNotFoundException e) {
+            throw new ResourceNotFoundException(e.getMessage());
+        }
     }
 
     //Método DELETE para deletar uma especie
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        especieRepository.deleteById(id);
+        if (!especieRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            especieRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DataBaseException("Falha na integridade referencial");
+        }
     }
 
     public void copyDtoToEntity(EspecieDTO dto, Especie entity) {
